refactor: 스탬프 컨트롤러 #64
